Fixed a warning.
git-svn-id: svn://svn.code.sf.net/p/dasm-dillon/code/trunk@209 77a3a24b-2c49-0410-90f5-8a0d088e77bf

+++ src/main.c
@@ -235,7 +235,7 @@ static void ShowSymbols( FILE *file, bool bTableSort )
         {
             /* Copy the element pointers into the symbol array */
             
-            int nPtr = 0;
+            size_t nPtr = 0;
             
             for (i = 0; i < SHASHSIZE; ++i)
                 for (sym = SHash[i]; sym; sym = sym->next)
